Refactor(jiwon/mypageAPI): review 관련 API user에서 분리 #69
